+++ src/models/Workspace.ts
@@ -37,6 +37,15 @@ export interface Invitations {
   createdAt: string;
 }
 
+export interface SentInvitation {
+  id: number;
+  workspaceId: number;
+  workspaceName: string;
+  invitedUserEmail: string;
+  status: string;
+  createdAt: string;
+}
+
 export interface ConfirmInvitations {
   status: boolean;
   id: number;